+++ tests/testthat/test-external-generic.R
@@ -1,4 +1,5 @@
 test_that("can get and append methods", {
+  external_methods_reset("S7")
   on.exit(external_methods_reset("S7"), add = TRUE)
 
   expect_equal(external_methods_get("S7"), list())
@@ -19,28 +20,34 @@ test_that("can get and append methods", {
 
 test_that("displays nicely", {
   bar <- new_external_generic("foo", "bar", "x")
+  on.exit(external_methods_reset("S7"), add = TRUE)
+
   expect_snapshot({
     print(bar)
   })
 })
 
 test_that("new_method works with both hard and soft dependencies", {
-  skip_on_os("windows")
+  # NB: Relies on installed S7
+
+  skip_if(getRversion() < 4.1 && Sys.info()[["sysname"]] == "Windows")
   skip_if(quick_test())
 
   tmp_lib <- tempfile()
   dir.create(tmp_lib)
   old_libpaths <- .libPaths()
   .libPaths(c(tmp_lib, old_libpaths))
+
   on.exit({
     .libPaths(old_libpaths)
-    detach("package:t2", unload = TRUE)
-    detach("package:t1", unload = TRUE)
-    detach("package:t0", unload = TRUE)
+    try(detach("package:t2", unload = TRUE), silent = TRUE)
+    try(detach("package:t1", unload = TRUE), silent = TRUE)
+    try(detach("package:t0", unload = TRUE), silent = TRUE)
     unlink(tmp_lib, recursive = TRUE)
   })
 
-  quick_install(test_path(c("t0", "t1", "t2")))
+  quick_install(test_path(c("t0", "t1")), tmp_lib)
+  quick_install(test_path("t2"), tmp_lib)
   library("t2")
 
   # t2 has a soft dependency on t1